Moving operations code to a new file
